Slim down to one plugin file
Similar to tests, most plugin devs probably won't ever bother splitting
their code into multiple files. Those who do are probably used to it,
and don't need our help figuring out the file structure.

Cutting down to a single file *might* also make reloading packaged
plugins work better until a proper solution can be implemented in Sopel
itself (but *if* it does, that's merely a side benefit of this change).
