+++ {{cookiecutter.repo_name}}/sopel_{{cookiecutter.module_name}}/__init__.py
@@ -5,9 +5,17 @@
 """
 from __future__ import unicode_literals, absolute_import, division, print_function
 
-from .{{cookiecutter.module_name}} import *
+from sopel import module
 
-__author__ = '{{cookiecutter.full_name}}'
-__email__ = '{{cookiecutter.email}}'
-__version__ = '{{cookiecutter.version}}'
 
+def configure(config):
+    pass
+
+
+def setup(bot):
+    pass
+
+
+@module.commands('helloworld')
+def hello_world(bot, trigger):
+    bot.say('Hello, world!')
